requirements.txt
streamlit>=1.35.0
streamlit-drawable-canvas>=0.9.3
tensorflow>=2.15.0
numpy>=1.24.0
pillow>=10.0.0
